requirements.txt
pycoingecko
schedule
flask_sqlalchemy
psycopg2-binary
flask
python-dotenv
flask-jwt-extended
gunicorn
flask-cors
pystan
fbprophet
yfinance
pandas
numpy
plotly


